requirements.txt
alembic==1.9.1
Bootstrap-Flask==2.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
flask-change-password==0.0.7
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mysql-connector==2.2.9
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.45
urllib3==1.26.13
Werkzeug==2.2.2
WTForms==3.0.1
